refactor: 전반적인 코드 리팩토링
- 기능은 이전과 같습니다
