--- src/main/java/com/example/emergencyassistb4b4/auth/dto/CreateAccessTokenRequestDto.java
@@ -1,10 +1,0 @@
-package com.example.emergencyassistb4b4.auth.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CreateAccessTokenRequestDto {
-    private String refreshToken;
-}
