+++ src/main/java/com/example/emergencyassistb4b4/auth/dto/TokenReissueRequestDto.java
@@ -1,0 +1,12 @@
+package com.example.emergencyassistb4b4.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class TokenReissueRequestDto {
+    @NotBlank(message = "리프레시 토큰은 필수입니다.")
+    private String refreshToken;
+}
